software/raspberry/requirements.txt
# Python dependencies for Raspberry Pi (rescue + line)

opencv-contrib-python
numpy
pyserial
ultralytics
onnxruntime

